[9주차/미카엘] 워크북 제출합니다

+++ src/main/java/umc/domain/mission/repository/MemberMissionRepository.java
@@ -3,5 +3,5 @@ package umc.domain.mission.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.domain.mission.entity.mapping.MemberMission;
 
-public interface MemberMissionRepository extends JpaRepository<MemberMission, Integer> , MemberMissionRepositoryCustom {
+public interface MemberMissionRepository extends JpaRepository<MemberMission, Long>, MemberMissionRepositoryCustom {
 }
